refactor: optimizer tab controller
