+++ src/lib/optimization/defaultForm.ts
@@ -132,45 +132,45 @@ export function defaultEnemyOptions() {
 }
 
 export const defaultSetConditionals = {
-  [Constants.Sets.PasserbyOfWanderingCloud]: [undefined, true],
-  [Constants.Sets.MusketeerOfWildWheat]: [undefined, true],
-  [Constants.Sets.KnightOfPurityPalace]: [undefined, true],
-  [Constants.Sets.HunterOfGlacialForest]: [undefined, true],
-  [Constants.Sets.ChampionOfStreetwiseBoxing]: [undefined, 5],
-  [Constants.Sets.GuardOfWutheringSnow]: [undefined, true],
-  [Constants.Sets.FiresmithOfLavaForging]: [undefined, true],
-  [Constants.Sets.GeniusOfBrilliantStars]: [undefined, true],
-  [Constants.Sets.BandOfSizzlingThunder]: [undefined, true],
-  [Constants.Sets.EagleOfTwilightLine]: [undefined, true],
-  [Constants.Sets.ThiefOfShootingMeteor]: [undefined, true],
-  [Constants.Sets.WastelanderOfBanditryDesert]: [undefined, 1],
-  [Constants.Sets.LongevousDisciple]: [undefined, 2],
-  [Constants.Sets.MessengerTraversingHackerspace]: [undefined, false],
-  [Constants.Sets.TheAshblazingGrandDuke]: [undefined, 0],
-  [Constants.Sets.PrisonerInDeepConfinement]: [undefined, 0],
-  [Constants.Sets.PioneerDiverOfDeadWaters]: [undefined, 2],
-  [Constants.Sets.WatchmakerMasterOfDreamMachinations]: [undefined, false],
-  [Constants.Sets.IronCavalryAgainstTheScourge]: [undefined, true],
-  [Constants.Sets.TheWindSoaringValorous]: [undefined, false],
-  [Constants.Sets.SacerdosRelivedOrdeal]: [undefined, 0],
-  [Constants.Sets.ScholarLostInErudition]: [undefined, true],
+  [Sets.PasserbyOfWanderingCloud]: [undefined, true],
+  [Sets.MusketeerOfWildWheat]: [undefined, true],
+  [Sets.KnightOfPurityPalace]: [undefined, true],
+  [Sets.HunterOfGlacialForest]: [undefined, true],
+  [Sets.ChampionOfStreetwiseBoxing]: [undefined, 5],
+  [Sets.GuardOfWutheringSnow]: [undefined, true],
+  [Sets.FiresmithOfLavaForging]: [undefined, true],
+  [Sets.GeniusOfBrilliantStars]: [undefined, true],
+  [Sets.BandOfSizzlingThunder]: [undefined, true],
+  [Sets.EagleOfTwilightLine]: [undefined, true],
+  [Sets.ThiefOfShootingMeteor]: [undefined, true],
+  [Sets.WastelanderOfBanditryDesert]: [undefined, 1],
+  [Sets.LongevousDisciple]: [undefined, 2],
+  [Sets.MessengerTraversingHackerspace]: [undefined, false],
+  [Sets.TheAshblazingGrandDuke]: [undefined, 0],
+  [Sets.PrisonerInDeepConfinement]: [undefined, 0],
+  [Sets.PioneerDiverOfDeadWaters]: [undefined, 2],
+  [Sets.WatchmakerMasterOfDreamMachinations]: [undefined, false],
+  [Sets.IronCavalryAgainstTheScourge]: [undefined, true],
+  [Sets.TheWindSoaringValorous]: [undefined, false],
+  [Sets.SacerdosRelivedOrdeal]: [undefined, 0],
+  [Sets.ScholarLostInErudition]: [undefined, true],
 
-  [Constants.Sets.SpaceSealingStation]: [undefined, true],
-  [Constants.Sets.FleetOfTheAgeless]: [undefined, true],
-  [Constants.Sets.PanCosmicCommercialEnterprise]: [undefined, true],
-  [Constants.Sets.BelobogOfTheArchitects]: [undefined, true],
-  [Constants.Sets.CelestialDifferentiator]: [undefined, false],
-  [Constants.Sets.InertSalsotto]: [undefined, true],
-  [Constants.Sets.TaliaKingdomOfBanditry]: [undefined, true],
-  [Constants.Sets.SprightlyVonwacq]: [undefined, true],
-  [Constants.Sets.RutilantArena]: [undefined, true],
-  [Constants.Sets.BrokenKeel]: [undefined, true],
-  [Constants.Sets.FirmamentFrontlineGlamoth]: [undefined, true],
-  [Constants.Sets.PenaconyLandOfTheDreams]: [undefined, true],
-  [Constants.Sets.SigoniaTheUnclaimedDesolation]: [undefined, 4],
-  [Constants.Sets.IzumoGenseiAndTakamaDivineRealm]: [undefined, true],
-  [Constants.Sets.DuranDynastyOfRunningWolves]: [undefined, 5],
-  [Constants.Sets.ForgeOfTheKalpagniLantern]: [undefined, false],
-  [Constants.Sets.LushakaTheSunkenSeas]: [undefined, false],
-  [Constants.Sets.TheWondrousBananAmusementPark]: [undefined, false],
+  [Sets.SpaceSealingStation]: [undefined, true],
+  [Sets.FleetOfTheAgeless]: [undefined, true],
+  [Sets.PanCosmicCommercialEnterprise]: [undefined, true],
+  [Sets.BelobogOfTheArchitects]: [undefined, true],
+  [Sets.CelestialDifferentiator]: [undefined, false],
+  [Sets.InertSalsotto]: [undefined, true],
+  [Sets.TaliaKingdomOfBanditry]: [undefined, true],
+  [Sets.SprightlyVonwacq]: [undefined, true],
+  [Sets.RutilantArena]: [undefined, true],
+  [Sets.BrokenKeel]: [undefined, true],
+  [Sets.FirmamentFrontlineGlamoth]: [undefined, true],
+  [Sets.PenaconyLandOfTheDreams]: [undefined, true],
+  [Sets.SigoniaTheUnclaimedDesolation]: [undefined, 4],
+  [Sets.IzumoGenseiAndTakamaDivineRealm]: [undefined, true],
+  [Sets.DuranDynastyOfRunningWolves]: [undefined, 5],
+  [Sets.ForgeOfTheKalpagniLantern]: [undefined, false],
+  [Sets.LushakaTheSunkenSeas]: [undefined, false],
+  [Sets.TheWondrousBananAmusementPark]: [undefined, false],
 }
