+++ src/lib/tabs/tabOptimizer/optimizerTabController.ts
@@ -1,30 +1,38 @@
 import { CellClickedEvent, IGetRowsParams, IRowNode } from 'ag-grid-community'
 import { inPlaceSort } from 'fast-sort'
-import { CharacterConditionalsResolver } from 'lib/conditionals/resolver/characterConditionalsResolver'
-import { LightConeConditionalsResolver } from 'lib/conditionals/resolver/lightConeConditionalsResolver'
-import { CombatBuffs, ConditionalDataType, Constants, DEFAULT_STAT_DISPLAY, Stats } from 'lib/constants/constants'
+import { Constants, DEFAULT_STAT_DISPLAY, Stats } from 'lib/constants/constants'
 import { SavedSessionKeys } from 'lib/constants/constantsSession'
 import { RelicsByPart } from 'lib/gpu/webgpuTypes'
 import { Message } from 'lib/interactions/message'
 import { OptimizerDisplayData, OptimizerDisplayDataStatSim } from 'lib/optimization/bufferPacker'
-import { defaultEnemyOptions, defaultSetConditionals, defaultTeammate, getDefaultForm, getDefaultWeights } from 'lib/optimization/defaultForm'
-import { ConditionalSetMetadata } from 'lib/optimization/rotation/setConditionalContent'
+import { getDefaultForm } from 'lib/optimization/defaultForm'
 import { GridAggregations } from 'lib/rendering/gradient'
 import { CharacterStats } from 'lib/scoring/characterStats'
 import DB from 'lib/state/db'
 import { SaveState } from 'lib/state/saveState'
 import { initializeComboState } from 'lib/tabs/tabOptimizer/combo/comboDrawerController'
-import { applyMetadataPresetToForm } from 'lib/tabs/tabOptimizer/optimizerForm/components/RecommendedPresetsButton'
+import { displayToForm, formToDisplay } from 'lib/tabs/tabOptimizer/optimizerForm/optimizerFormTransform'
+import { optimizerGridApi } from 'lib/utils/gridUtils'
 import { TsUtils } from 'lib/utils/TsUtils'
-import { Utils } from 'lib/utils/utils'
-import { Form, Teammate } from 'types/form'
-import { OptimizerCombatBuffs } from 'types/optimizer'
+import { Form } from 'types/form'
 
-// FIXME HIGH
+type PermutationSizes = {
+  hSize: number
+  gSize: number
+  bSize: number
+  fSize: number
+  pSize: number
+  lSize: number
+}
+
+type SortModel = {
+  colId: string
+  sort: string
+}
 
 let relics: RelicsByPart
-let consts: OptimizerSizes
-let aggs: GridAggregations
+let permutationSizes: PermutationSizes
+let aggregations: GridAggregations
 let rows: OptimizerDisplayData[] = []
 let filteredIndices: number[]
 let filterModel: Form
@@ -42,7 +50,9 @@ const columnsToAggregateMap = {
   [Stats.BE]: true,
   [Stats.ERR]: true,
   [Stats.OHB]: true,
-  // For custom ones remember to set the min/max in aggregate()
+
+  // For custom rows remember to set the min/max in aggregate()
+
   ED: true,
   WEIGHT: true,
   EHP: true,
@@ -72,46 +82,28 @@ const columnsToAggregateMap = {
 }
 const columnsToAggregate = Object.keys(columnsToAggregateMap)
 
-type OptimizerSizes = {
-  hSize: number
-  gSize: number
-  bSize: number
-  fSize: number
-  pSize: number
-  lSize: number
-}
-
-type SortModel = {
-  colId: string
-  sort: string
-}
-
 export const OptimizerTabController = {
-  setMetadata: (inputConsts: OptimizerSizes, inputRelics: RelicsByPart) => {
-    consts = inputConsts
+  setMetadata: (inputPermutationSizes: PermutationSizes, inputRelics: RelicsByPart) => {
+    permutationSizes = inputPermutationSizes
     relics = inputRelics
   },
 
-  setAggs: (x: GridAggregations) => {
-    aggs = x
+  getAggregations: () => {
+    return aggregations
   },
 
-  getAggs: () => {
-    return aggs
+  setRows: (newRows: OptimizerDisplayData[]) => {
+    rows = newRows
   },
 
-  setRows: (x: OptimizerDisplayData[]) => {
-    rows = x
-  },
-
-  setTopRow: (x: OptimizerDisplayData, overwrite = false) => {
+  setTopRow: (row: OptimizerDisplayData, overwrite = false) => {
     if (overwrite) {
-      window.optimizerGrid.current?.api.updateGridOptions({ pinnedTopRowData: [x] })
+      window.optimizerGrid.current?.api.updateGridOptions({ pinnedTopRowData: [row] })
       return
     }
 
     const currentPinned = window.optimizerGrid.current?.api.getGridOption('pinnedTopRowData') ?? []
-    currentPinned[0] = x
+    currentPinned[0] = row
     window.optimizerGrid.current?.api.updateGridOptions({ pinnedTopRowData: currentPinned })
   },
 
@@ -123,18 +115,36 @@ export const OptimizerTabController = {
     document.getElementById('optimizerGridContainer')!.scrollIntoView({ behavior: 'smooth', block: 'nearest' })
   },
 
+  // Get a form that's ready for optimizer submission
+  getForm: () => {
+    const form = window.optimizerForm.getFieldsValue() as Form
+    return OptimizerTabController.displayToForm(form)
+  },
+
+  // Convert a form to its visual representation
+  formToDisplay: (form: Form) => {
+    return formToDisplay(form)
+  },
+
+  // Parse out any invalid values and prepare the form for submission to optimizer
+  displayToForm: (form: Form) => {
+    return displayToForm(form)
+  },
+
   equipClicked: () => {
     console.log('Equip clicked')
-    const formValues = OptimizerTabController.getForm()
-    const characterId = formValues.characterId
+    const form = OptimizerTabController.getForm()
+    const characterId = form.characterId
 
     if (!characterId) {
       return
     }
-    DB.addFromForm(formValues)
+
+    DB.addFromForm(form)
 
     const selectedNodes = window.optimizerGrid.current?.api.getSelectedNodes() as IRowNode<OptimizerDisplayDataStatSim>[]
     if (!selectedNodes || selectedNodes.length == 0 || (selectedNodes[0]?.data?.statSim)) {
+      // Cannot equip a stat sim or empty row
       return
     }
 
@@ -151,16 +161,18 @@ export const OptimizerTabController = {
 
   cellClicked: (event: CellClickedEvent) => {
     const data = event.data as OptimizerDisplayDataStatSim
+    const gridApi = optimizerGridApi()
 
     if (event.rowPinned == 'top') {
-      console.log('Top row clicked', event.data)
-      const fieldValues = OptimizerTabController.getForm()
-      if (event.data && fieldValues.characterId) {
-        const character = DB.getCharacterById(fieldValues.characterId)
-
+      // Clicking the top row should display current relics
+      console.log('Top row clicked', data)
+      const form = OptimizerTabController.getForm()
+      if (data && form.characterId) {
         if (!data.id) {
-          window.optimizerGrid.current?.api.deselectAll()
+          gridApi.deselectAll()
         }
+
+        const character = DB.getCharacterById(form.characterId)
 
         if (character && data.id) {
           const rowId = data.id
@@ -168,9 +180,9 @@ export const OptimizerTabController = {
           window.store.getState().setOptimizerBuild(build)
 
           // Find the row by its string ID and select it
-          const rowNode: IRowNode<OptimizerDisplayData> = window.optimizerGrid.current?.api.getRowNode(String(data.id)) as IRowNode<OptimizerDisplayData>
+          const rowNode: IRowNode<OptimizerDisplayData> = gridApi.getRowNode(String(data.id)) as IRowNode<OptimizerDisplayData>
           if (rowNode) {
-            const currentPinned: OptimizerDisplayData[] = window.optimizerGrid.current?.api.getGridOption('pinnedTopRowData') ?? []
+            const currentPinned: OptimizerDisplayData[] = gridApi.getGridOption('pinnedTopRowData') ?? []
 
             if (String(currentPinned[0].id) == String(rowNode.data!.id)) {
               // The currently equipped top row shouldn't correspond to an optimizer row, deselect
@@ -219,7 +231,7 @@ export const OptimizerTabController = {
     return {
       getRows: (params: IGetRowsParams) => {
         // @ts-ignore
-        aggs = undefined
+        aggregations = undefined
 
         // fast clickers can race unmount/remount and cause NPE here.
         if (window?.optimizerGrid?.current?.api) {
@@ -259,13 +271,14 @@ export const OptimizerTabController = {
     }
   },
 
+  // Unpack a permutation ID to its respective relics
   calculateRelicsFromId: (id: number) => {
-    const lSize = consts.lSize
-    const pSize = consts.pSize
-    const fSize = consts.fSize
-    const bSize = consts.bSize
-    const gSize = consts.gSize
-    const hSize = consts.hSize
+    const lSize = permutationSizes.lSize
+    const pSize = permutationSizes.pSize
+    const fSize = permutationSizes.fSize
+    const bSize = permutationSizes.bSize
+    const gSize = permutationSizes.gSize
+    const hSize = permutationSizes.hSize
 
     const x = id
     const l = (x % lSize)
@@ -297,249 +310,6 @@ export const OptimizerTabController = {
       LinkRope: relicsFromId.LinkRope.id,
     }
   },
-  // Get a form that's ready for submission
-  getForm: () => {
-    const form = window.optimizerForm.getFieldsValue() as Form
-    return OptimizerTabController.fixForm(form)
-  },
-
-  // Convert a form to its visual representation
-  getDisplayFormValues: (form: Form) => {
-    const characterId = form.characterId
-    const newForm: Partial<Form> = TsUtils.clone(form)
-    const metadata = DB.getMetadata().characters[characterId]
-    const scoringMetadata = DB.getScoringMetadata(characterId)
-
-    // Erase inputs where min == 0 and max == MAX_INT
-    newForm.maxHp = unsetMax(form.maxHp)
-    newForm.minHp = unsetMin(form.minHp)
-    newForm.maxAtk = unsetMax(form.maxAtk)
-    newForm.minAtk = unsetMin(form.minAtk)
-    newForm.maxDef = unsetMax(form.maxDef)
-    newForm.minDef = unsetMin(form.minDef)
-    newForm.maxSpd = unsetMax(form.maxSpd)
-    newForm.minSpd = unsetMin(form.minSpd)
-    newForm.maxCr = unsetMax(form.maxCr, true)
-    newForm.minCr = unsetMin(form.minCr, true)
-    newForm.maxCd = unsetMax(form.maxCd, true)
-    newForm.minCd = unsetMin(form.minCd, true)
-    newForm.maxEhr = unsetMax(form.maxEhr, true)
-    newForm.minEhr = unsetMin(form.minEhr, true)
-    newForm.maxRes = unsetMax(form.maxRes, true)
-    newForm.minRes = unsetMin(form.minRes, true)
-    newForm.maxBe = unsetMax(form.maxBe, true)
-    newForm.minBe = unsetMin(form.minBe, true)
-    newForm.maxErr = unsetMax(form.maxErr, true)
-    newForm.minErr = unsetMin(form.minErr, true)
-
-    newForm.maxEhp = unsetMax(form.maxEhp)
-    newForm.minEhp = unsetMin(form.minEhp)
-    newForm.maxBasic = unsetMax(form.maxBasic)
-    newForm.minBasic = unsetMin(form.minBasic)
-    newForm.maxSkill = unsetMax(form.maxSkill)
-    newForm.minSkill = unsetMin(form.minSkill)
-    newForm.maxUlt = unsetMax(form.maxUlt)
-    newForm.minUlt = unsetMin(form.minUlt)
-    newForm.maxFua = unsetMax(form.maxFua)
-    newForm.minFua = unsetMin(form.minFua)
-    newForm.maxDot = unsetMax(form.maxDot)
-    newForm.minDot = unsetMin(form.minDot)
-    newForm.maxBreak = unsetMax(form.maxBreak)
-    newForm.minBreak = unsetMin(form.minBreak)
-    newForm.maxHeal = unsetMax(form.maxHeal)
-    newForm.minHeal = unsetMin(form.minHeal)
-    newForm.maxShield = unsetMax(form.maxShield)
-    newForm.minShield = unsetMin(form.minShield)
-
-    const combatBuffs = {} as Partial<OptimizerCombatBuffs>
-    if (!form.combatBuffs) form.combatBuffs = {}
-    for (const buff of Object.values(CombatBuffs)) {
-      combatBuffs[buff.key as keyof OptimizerCombatBuffs] = unsetMin(form.combatBuffs[buff.key], buff.percent)
-    }
-    newForm.combatBuffs = combatBuffs
-
-    if (!newForm.setConditionals) {
-      newForm.setConditionals = defaultSetConditionals
-    } else {
-      Utils.mergeUndefinedValues(newForm.setConditionals, defaultSetConditionals)
-    }
-
-    const enemyOptions = defaultEnemyOptions()
-
-    if (!form.enemyLevel) {
-      newForm.enemyLevel = enemyOptions.enemyLevel
-    }
-
-    if (!form.enemyCount) {
-      newForm.enemyCount = enemyOptions.enemyCount
-    }
-
-    if (!form.enemyResistance) {
-      newForm.enemyResistance = enemyOptions.enemyResistance
-    }
-
-    if (!form.enemyEffectResistance) {
-      newForm.enemyEffectResistance = enemyOptions.enemyEffectResistance
-    }
-
-    if (form.enemyElementalWeak == null) {
-      newForm.enemyElementalWeak = enemyOptions.enemyElementalWeak
-    }
-
-    if (form.enemyWeaknessBroken == null) {
-      newForm.enemyWeaknessBroken = enemyOptions.enemyWeaknessBroken
-    }
-
-    if (!form.enemyMaxToughness) {
-      newForm.enemyMaxToughness = enemyOptions.enemyMaxToughness
-    }
-
-    if (newForm.characterId) {
-      const defaultOptions = CharacterConditionalsResolver.get(form).defaults()
-      if (!newForm.characterConditionals) {
-        newForm.characterConditionals = {}
-      }
-      for (const option of Object.keys(defaultOptions)) {
-        if (newForm.characterConditionals[option] == undefined) {
-          newForm.characterConditionals[option] = defaultOptions[option]
-        }
-      }
-    }
-
-    if (newForm.lightCone) {
-      const defaultLcOptions = LightConeConditionalsResolver.get(form).defaults()
-      if (!newForm.lightConeConditionals) {
-        newForm.lightConeConditionals = {}
-      }
-      for (const option of Object.keys(defaultLcOptions)) {
-        if (newForm.lightConeConditionals[option] == undefined) {
-          newForm.lightConeConditionals[option] = defaultLcOptions[option]
-        }
-      }
-    } else {
-      newForm.lightCone = undefined
-      newForm.lightConeLevel = 80
-      newForm.lightConeSuperimposition = 1
-      newForm.lightConeConditionals = {}
-    }
-
-    if (!newForm.statDisplay) {
-      newForm.statDisplay = DEFAULT_STAT_DISPLAY
-    }
-
-    const character = DB.getCharacterById(characterId)
-    if (character) {
-      newForm.rank = character.rank
-    } else {
-      newForm.rank = DB.getCharacters().length
-
-      // Apply any presets to new characters
-      if (metadata) {
-        for (const preset of scoringMetadata.presets || []) {
-          preset.apply(newForm)
-        }
-
-        newForm.mainBody = scoringMetadata.parts[Constants.Parts.Body]
-        newForm.mainFeet = scoringMetadata.parts[Constants.Parts.Feet]
-        newForm.mainPlanarSphere = scoringMetadata.parts[Constants.Parts.PlanarSphere]
-        newForm.mainLinkRope = scoringMetadata.parts[Constants.Parts.LinkRope]
-        newForm.weights = scoringMetadata.stats
-        newForm.weights.headHands = 0
-        newForm.weights.bodyFeet = 0
-        newForm.weights.sphereRope = 0
-
-        applyMetadataPresetToForm(newForm as Form, scoringMetadata)
-      }
-    }
-
-    if (!newForm.weights) {
-      newForm.weights = getDefaultWeights(characterId)
-    }
-
-    if (!newForm.weights.headHands) {
-      newForm.weights.headHands = 0
-    }
-    if (!newForm.weights.bodyFeet) {
-      newForm.weights.bodyFeet = 0
-    }
-    if (!newForm.weights.sphereRope) {
-      newForm.weights.sphereRope = 0
-    }
-
-    if (!newForm.exclude) {
-      newForm.exclude = []
-    }
-
-    // Clean up any deleted excluded characters
-    newForm.exclude = newForm.exclude.filter((id) => DB.getCharacterById(id))
-
-    // Some pre-existing saves had this default to undefined while the toggle defaults to true and hides the undefined.
-    // Keeping this here for now
-    if (newForm.includeEquippedRelics == null) {
-      newForm.includeEquippedRelics = true
-    }
-
-    if (![1, 3, 5].includes(newForm.enemyCount!)) {
-      newForm.enemyCount = 1
-    }
-
-    newForm.teammate0 = cloneTeammate(form.teammate0)
-    newForm.teammate1 = cloneTeammate(form.teammate1)
-    newForm.teammate2 = cloneTeammate(form.teammate2)
-
-    if (!newForm.weights) {
-      newForm.weights = getDefaultWeights()
-    }
-
-    if (!newForm.resultSort) {
-      if (metadata) {
-        newForm.resultSort = metadata.scoringMetadata.sortOption.key
-      }
-    }
-
-    if (!newForm.resultsLimit) {
-      newForm.resultsLimit = 1024
-    }
-
-    if (!newForm.mainStatUpscaleLevel) {
-      newForm.mainStatUpscaleLevel = 15
-    }
-
-    if (!newForm.comboStateJson) {
-      newForm.comboStateJson = '{}'
-    }
-
-    if (!newForm.comboAbilities) {
-      const simulation = metadata.scoringMetadata?.simulation
-      newForm.comboAbilities = simulation?.comboAbilities ?? [null, 'BASIC']
-      newForm.comboDot = simulation?.comboDot ?? 0
-      newForm.comboBreak = simulation?.comboBreak ?? 0
-    }
-
-    if (!newForm.comboType) {
-      newForm.comboType = 'simple'
-    }
-
-    for (const [key, value] of Object.entries(newForm.setConditionals)) {
-      if (!ConditionalSetMetadata[key]) {
-        delete form.setConditionals[key]
-        delete newForm.setConditionals[key]
-        continue
-      }
-      if (ConditionalSetMetadata[key].type === ConditionalDataType.SELECT) {
-        if (typeof value[1] != 'number') {
-          newForm.setConditionals[key][1] = defaultSetConditionals[key][1]
-        }
-      } else {
-        if (typeof value[1] != 'boolean') {
-          newForm.setConditionals[key][1] = defaultSetConditionals[key][1]
-        }
-      }
-    }
-
-    // console.log('Form update', newForm)
-    return newForm as Form
-  },
 
   validateForm: (form: Form) => {
     console.log('validate', form)
@@ -567,76 +337,16 @@ export const OptimizerTabController = {
       return false
     }
 
-    const lcMeta = DB.getMetadata().lightCones[form.lightCone]
-    const charMeta = DB.getMetadata().characters[form.characterId]
+    const metadata = DB.getMetadata()
+    const lcMeta = metadata.lightCones[form.lightCone]
+    const charMeta = metadata.characters[form.characterId]
+
     if (lcMeta.path != charMeta.path) {
       Message.warning('Character path doesn\'t match light cone path.', 10)
       console.log('Path mismatch')
     }
 
     return true
-  },
-
-  // Parse out any invalid values and prepare the form for submission to optimizer
-  fixForm: (form: Form) => {
-    const MAX_INT = Constants.MAX_INT
-
-    form.statDisplay = window.store.getState().statDisplay || DEFAULT_STAT_DISPLAY
-
-    form.maxHp = fixValue(form.maxHp, MAX_INT)
-    form.minHp = fixValue(form.minHp, 0)
-    form.maxAtk = fixValue(form.maxAtk, MAX_INT)
-    form.minAtk = fixValue(form.minAtk, 0)
-    form.maxDef = fixValue(form.maxDef, MAX_INT)
-    form.minDef = fixValue(form.minDef, 0)
-    form.maxSpd = fixValue(form.maxSpd, MAX_INT)
-    form.minSpd = fixValue(form.minSpd, 0)
-    form.maxCr = fixValue(form.maxCr, MAX_INT, 100)
-    form.minCr = fixValue(form.minCr, 0, 100)
-    form.maxCd = fixValue(form.maxCd, MAX_INT, 100)
-    form.minCd = fixValue(form.minCd, 0, 100)
-    form.maxEhr = fixValue(form.maxEhr, MAX_INT, 100)
-    form.minEhr = fixValue(form.minEhr, 0, 100)
-    form.maxRes = fixValue(form.maxRes, MAX_INT, 100)
-    form.minRes = fixValue(form.minRes, 0, 100)
-    form.maxBe = fixValue(form.maxBe, MAX_INT, 100)
-    form.minBe = fixValue(form.minBe, 0, 100)
-    form.maxErr = fixValue(form.maxErr, MAX_INT, 100)
-    form.minErr = fixValue(form.minErr, 0, 100)
-
-    form.maxEhp = fixValue(form.maxEhp, MAX_INT)
-    form.minEhp = fixValue(form.minEhp, 0)
-
-    form.maxBasic = fixValue(form.maxBasic, MAX_INT)
-    form.minBasic = fixValue(form.minBasic, 0)
-    form.maxSkill = fixValue(form.maxSkill, MAX_INT)
-    form.minSkill = fixValue(form.minSkill, 0)
-    form.maxUlt = fixValue(form.maxUlt, MAX_INT)
-    form.minUlt = fixValue(form.minUlt, 0)
-    form.maxFua = fixValue(form.maxFua, MAX_INT)
-    form.minFua = fixValue(form.minFua, 0)
-    form.maxDot = fixValue(form.maxDot, MAX_INT)
-    form.minDot = fixValue(form.minDot, 0)
-    form.maxBreak = fixValue(form.maxBreak, MAX_INT)
-    form.minBreak = fixValue(form.minBreak, 0)
-    form.maxHeal = fixValue(form.maxHeal, MAX_INT)
-    form.minHeal = fixValue(form.minHeal, 0)
-    form.maxShield = fixValue(form.maxShield, MAX_INT)
-    form.minShield = fixValue(form.minShield, 0)
-
-    if (!form.combatBuffs) form.combatBuffs = {}
-    for (const buff of Object.values(CombatBuffs)) {
-      form.combatBuffs[buff.key] = fixValue(form.combatBuffs[buff.key], 0, buff.percent ? 100 : 0)
-    }
-
-    form.mainHead = form.mainHead || []
-    form.mainHands = form.mainHands || []
-    form.mainBody = form.mainBody || []
-    form.mainFeet = form.mainFeet || []
-    form.mainPlanarSphere = form.mainPlanarSphere || []
-    form.mainLinkRope = form.mainLinkRope || []
-
-    return form
   },
 
   updateFilters: () => {
@@ -647,7 +357,6 @@ export const OptimizerTabController = {
   },
 
   resetFilters: () => {
-    console.info('@resetFilters')
     const fieldValues = OptimizerTabController.getForm()
     const newForm: Partial<Form> = {
       characterEidolon: fieldValues.characterEidolon,
@@ -672,7 +381,7 @@ export const OptimizerTabController = {
       relicSets: [],
     }
 
-    window.optimizerForm.setFieldsValue(OptimizerTabController.getDisplayFormValues(newForm as Form))
+    window.optimizerForm.setFieldsValue(OptimizerTabController.formToDisplay(newForm as Form))
     OptimizerTabController.updateFilters()
   },
 
@@ -692,7 +401,7 @@ export const OptimizerTabController = {
     const character = DB.getCharacterById(characterId)
 
     const form = character ? character.form : getDefaultForm({ id: characterId })
-    const displayFormValues = OptimizerTabController.getDisplayFormValues(form as Form)
+    const displayFormValues = OptimizerTabController.formToDisplay(form as Form)
     window.optimizerForm.setFieldsValue(displayFormValues)
 
     const comboState = initializeComboState(displayFormValues, true)
@@ -723,24 +432,6 @@ export const OptimizerTabController = {
     console.log('Apply filters to rows', fieldValues)
     OptimizerTabController.resetDataSource()
   },
-}
-
-function unsetMin(value: number, percent: boolean = false) {
-  if (value == undefined) return undefined
-  return value == 0 ? undefined : parseFloat((percent ? value * 100 : value).toFixed(3))
-}
-
-function unsetMax(value: number, percent: boolean = false) {
-  if (value == undefined) return undefined
-  return value == Constants.MAX_INT ? undefined : parseFloat((percent ? value * 100 : value).toFixed(3))
-}
-
-function fixValue(value: number, def: number, div: number = 0) {
-  if (value == null) {
-    return def
-  }
-  div = div || 1
-  return value / div
 }
 
 function aggregate(subArray: OptimizerDisplayData[]) {
@@ -791,7 +482,7 @@ function aggregate(subArray: OptimizerDisplayData[]) {
       if (value > maxAgg[column]) maxAgg[column] = value
     }
   }
-  aggs = {
+  aggregations = {
     minAgg: minAgg,
     maxAgg: maxAgg,
   }
@@ -866,14 +557,4 @@ function filter(filterModel: Form) {
   }
 
   filteredIndices = indices
-}
-
-function cloneTeammate(teammate: Teammate | undefined) {
-  if (!teammate?.characterId) return defaultTeammate() as Teammate
-  return {
-    characterId: teammate.characterId ?? null,
-    characterEidolon: teammate.characterEidolon ?? null,
-    lightCone: teammate.lightCone ?? null,
-    lightConeSuperimposition: teammate.lightConeSuperimposition ?? null,
-  } as Teammate
 }
