+++ src/lib/utils/gridUtils.ts
@@ -1,0 +1,4 @@
+export function optimizerGridApi() {
+  const grid = window.optimizerGrid.current!
+  return grid.api
+}
